implement zoom

+++ Assets/Scripts/OmniTwin/Core/OmniWorld.cs
@@ -6,6 +6,6 @@ namespace OmniTwin
     public static class OmniWorld
     {
         public static CesiumGlobeAnchor CesiumGlobeAnchor;
-        public static float3 CameraSpeed;
+        public static double3 CameraSpeed;
     }
 }
